immutable validation file and better creation

+++ pam_test.c
@@ -20,6 +20,8 @@
 #include <errno.h>
 #include <sys/mount.h>
 #include <sys/statfs.h>
+#include <sys/ioctl.h>
+#include <linux/fs.h>
 
 #include "utils/enc_utils.h"
 #include "argon2.h"
@@ -87,11 +89,21 @@ int create_and_encrypt_validation_file(char directory[PATH_MAX], char * content)
   char filename_enc[PATH_MAX];
   sprintf(filename_enc, "%s/.validation_file", directory);
 
+
+
+  int ret = access(directory, F_OK);
+  if (ret == 0) { // directory exists
+    if(access(directory, R_OK) == -1)
+      return 0; // PAM Module doesnt have read permissions, so the directory and validation file was already created in the past
+    
+    return 0;
+  }
+  if(ret == -1) {
+    perror("accessing directory");
+  }
   
-  int ret = access(filename_enc, F_OK);
-  printf("access(%s) = %d\n", filename_enc,ret);
+  ret = access(filename_enc, F_OK);
   if (ret == 0) {
-    printf("file existing\n");
     return 0;
   }
   if(ret == -1) {
@@ -107,9 +119,21 @@ int create_and_encrypt_validation_file(char directory[PATH_MAX], char * content)
   if (w == -1 || w != len_content) {
     return -1;
   }
+
   close(fd);
 
   encrypt_file(filename_dec, filename_enc, hash);
+
+  int fd2 = open(filename_enc, O_RDONLY);
+  if (fd == -1) {
+    return -1;
+  }
+
+  int immutable_flag = FS_IMMUTABLE_FL;
+  if (ioctl(fd2,FS_IOC_SETFLAGS, &immutable_flag) == -1) {
+    perror("ioctl");
+    return -1;
+  }
 
   unlink(filename_dec);
   return 0;
@@ -123,13 +147,13 @@ int pam_sm_authenticate(pam_handle_t *pamh, int flags, int argc, const char **ar
   const char *user;
   if (pam_get_user(pamh, &user, NULL) != PAM_SUCCESS)
   {
-      exit_pam("failed to get username\n");
-      return PAM_IGNORE;
+    exit_pam("failed to get username\n");
+    return PAM_IGNORE;
   }
   if ((pw = getpwnam(user)) == NULL)
   {
-      exit_pam("couldn't find username\n");
-      return PAM_IGNORE;
+    exit_pam("couldn't find username\n");
+    return PAM_IGNORE;
   }
   if (strcmp(pw->pw_name, "root") == 0 ){
     exit_pam("fuse not working for root\n");
@@ -148,20 +172,18 @@ int pam_sm_authenticate(pam_handle_t *pamh, int flags, int argc, const char **ar
   const char * password = NULL;
   pam_get_authtok(pamh, PAM_AUTHTOK, &password, NULL);
 
-
   // compute hash from user password
   compute_hash(password);
 
   // create directories to save data and to mount the filesystem on
-
   int s_dir = mkdir(dir_name, 0770);
   if ( s_dir == -1 && errno != EEXIST ) {
       exit_pam("failed to create mount directory\n");
       return PAM_IGNORE;
   }
-  else {
-    printf("private files will be visible in %s\n", dir_name);
-  }
+  // else {
+  //   printf("private files will be visible in %s\n", dir_name);
+  // }
 
   // create directory to save data in
   char data_dir_name[PATH_MAX];
@@ -172,17 +194,18 @@ int pam_sm_authenticate(pam_handle_t *pamh, int flags, int argc, const char **ar
       exit_pam("failed to create data directory\n");
       return PAM_IGNORE;
   }
-  else {
-    printf("private files will be saved in %s\n", data_dir_name);
-  }
+  // else {
+  //   printf("private files will be saved in %s\n", data_dir_name);
+  // }
 
-  printf("create and enc val file\n");
+  // create and encrypt the validation file in the .private directory
   if (create_and_encrypt_validation_file(data_dir_name, data_dir_name) != 0 ) {
     perror("");
     exit_pam("creating and encrypting validation file");
     return PAM_IGNORE;
   }
 
+  // change directories owner and group
   chown(dir_name, pw->pw_uid, pw->pw_gid);
   chown(data_dir_name, pw->pw_uid, pw->pw_gid);
